+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable end-user facing changes should be documented in this file.
 
+## [1.2.4] 2026-01-06
+
+### Fixed
+
+- There was a small bug in `ColorAndScaleTranstioner` where it was incorrectly canceling color/alpha tweens immediately
+  when there was a scale transition starting at the same time. This is now fixed. 
+
 
 ## [1.2.3] 2025-09-01